chore(deps): run React Compiler on `oxc-transform-react` instead of babel
Every React Compiler pass ran through babel: `@sanity/pkg-utils`
builds loaded `babel-plugin-react-compiler` via `reactCompiler`, and
the playground, example, and vitest configs stacked
`@rolldown/plugin-babel` + `reactCompilerPreset` on top of
`@vitejs/plugin-react`, a full babel parse/transform/codegen
round-trip per file on top of the oxc transform vite already runs.

`transform: 'oxc'` in every `reactCompiler` config and
`react({compiler: {target: '19'}})` in the vite/vitest configs now run
the compiler through `oxc-transform-react`, the native Rust port: one
pass handles React Compiler, TypeScript/JSX and Fast Refresh.
`babel-plugin-react-compiler` and `@rolldown/plugin-babel` leave every
manifest and the catalog; `oxc-transform-react` and its native
bindings join `minimumReleaseAgeExclude` (the release is days old).

oxc `sources` only accepts substring filters, so the configs that
excluded the engine by regex convert to include-lists:
`COMPILED_SOURCES` moves from `packages/editor/package.config.ts` to
`react-compiler-sources.ts` and is shared by the package build, the
editor vitest projects (widened with the `tests/`, `gherkin-tests/`
and `vitest.setup.ts` harnesses the old exclude let through), and the
playground (widened with its own `src/` and the aliased plugin
sources). The compiled set is otherwise unchanged: in the built lib,
tier-3 `element.tsx` still gets a memo cache and the untiered
`useChildren` still does not. Two things fall away: the playground's
`\0rolldown/runtime.js` re-exclusion (an include-list cannot match
it), and babel's silence about skipped functions (oxc prints
per-function bailout diagnostics).
