+++ apps/playground/package.json
@@ -28,7 +28,6 @@
     "@portabletext/plugin-typography": "workspace:*",
     "@portabletext/react": "catalog:tooling",
     "@portabletext/toolbar": "workspace:*",
-    "@rolldown/plugin-babel": "catalog:tooling",
     "@xstate/react": "catalog:",
     "emojilib": "^4.0.2",
     "fuse.js": "^7.1.0",
@@ -48,7 +47,7 @@
     "@types/react-dom": "catalog:tooling",
     "@types/react-is": "^19.2.0",
     "@vitejs/plugin-react": "catalog:tooling",
-    "babel-plugin-react-compiler": "catalog:tooling",
+    "oxc-transform-react": "catalog:tooling",
     "tailwind-merge": "catalog:tooling",
     "tailwind-variants": "^1.0.0",
     "tailwindcss": "catalog:tooling",
