+++ apps/playground/tsconfig.node.json
@@ -6,8 +6,12 @@
     "module": "ESNext",
     "moduleResolution": "bundler",
     "allowSyntheticDefaultImports": true,
+    "allowImportingTsExtensions": true,
     "strict": true,
     "noEmit": true
   },
-  "include": ["vite.config.ts"]
+  "include": [
+    "vite.config.ts",
+    "../../packages/editor/react-compiler-sources.ts"
+  ]
 }
