+++ examples/basic/vite.config.ts
@@ -1,11 +1,10 @@
 import path from 'node:path'
-import babel from '@rolldown/plugin-babel'
-import react, {reactCompilerPreset} from '@vitejs/plugin-react'
+import react from '@vitejs/plugin-react'
 import {defineConfig} from 'vite'
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [react(), babel({presets: [reactCompilerPreset({target: '19'})]})],
+  plugins: [react({compiler: {target: '19'}})],
   resolve: {
     alias: {
       '@portabletext/editor': path.resolve(
