+++ knip.json
@@ -6,7 +6,7 @@
     "packages/*": {
       "entry": ["src/index.ts!", "package.config.ts!"],
       "ignoreDependencies": [
-        "babel-plugin-react-compiler",
+        "oxc-transform-react",
         "@vitest/browser",
         "@vitest/browser-playwright",
         "@vitest/coverage-v8",
@@ -26,7 +26,7 @@
         "package.config.ts!"
       ],
       "ignoreDependencies": [
-        "babel-plugin-react-compiler",
+        "oxc-transform-react",
         "@vitest/browser",
         "@vitest/browser-playwright",
         "@vitest/coverage-v8",
@@ -42,7 +42,7 @@
         "package.config.ts!"
       ],
       "ignoreDependencies": [
-        "babel-plugin-react-compiler",
+        "oxc-transform-react",
         "@vitest/browser",
         "@vitest/browser-playwright",
         "@vitest/coverage-v8",
@@ -52,7 +52,7 @@
     },
     "apps/*": {
       "ignoreDependencies": [
-        "babel-plugin-react-compiler",
+        "oxc-transform-react",
         "tailwindcss-react-aria-components",
         "sharp"
       ]
