+++ packages/plugin-dnd/package.json
@@ -55,7 +55,6 @@
     "@portabletext/editor": "workspace:^",
     "@portabletext/schema": "workspace:^",
     "@portabletext/test": "workspace:^",
-    "@rolldown/plugin-babel": "catalog:tooling",
     "@sanity/pkg-utils": "catalog:tooling",
     "@sanity/tsconfig": "catalog:tooling",
     "@types/react": "catalog:tooling",
@@ -63,7 +62,7 @@
     "@vitejs/plugin-react": "catalog:tooling",
     "@vitest/browser": "catalog:tooling",
     "@vitest/browser-playwright": "catalog:tooling",
-    "babel-plugin-react-compiler": "catalog:tooling",
+    "oxc-transform-react": "catalog:tooling",
     "react": "catalog:",
     "typescript": "catalog:tooling",
     "vite": "catalog:tooling",
