+++ packages/plugin-emoji-picker/vitest.config.ts
@@ -1,5 +1,4 @@
-import babel from '@rolldown/plugin-babel'
-import react, {reactCompilerPreset} from '@vitejs/plugin-react'
+import react from '@vitejs/plugin-react'
 import {playwright} from '@vitest/browser-playwright'
 import {defineConfig} from 'vitest/config'
 
@@ -10,10 +9,7 @@ export default defineConfig({
         // Resolves the monorepo-only `@portabletext/editor/test*` specifiers
         // through this package's tsconfig `paths`.
         resolve: {tsconfigPaths: true},
-        plugins: [
-          react(),
-          babel({presets: [reactCompilerPreset({target: '19'})]}),
-        ],
+        plugins: [react({compiler: {target: '19'}})],
         test: {
           name: 'browser',
           browser: {
