+++ packages/plugin-input-rule/package.config.ts
@@ -2,5 +2,5 @@ import {defineConfig} from '@sanity/pkg-utils'
 
 export default defineConfig({
   tsconfig: 'tsconfig.dist.json',
-  reactCompiler: {target: '19'},
+  reactCompiler: {transform: 'oxc', target: '19'},
 })
