+++ packages/plugin-paste-link/package.json
@@ -59,14 +59,13 @@
   },
   "devDependencies": {
     "@portabletext/editor": "workspace:^",
-    "@rolldown/plugin-babel": "catalog:tooling",
     "@sanity/pkg-utils": "catalog:tooling",
     "@sanity/tsconfig": "catalog:tooling",
     "@types/react": "catalog:tooling",
     "@vitejs/plugin-react": "catalog:tooling",
     "@vitest/browser": "catalog:tooling",
     "@vitest/browser-playwright": "catalog:tooling",
-    "babel-plugin-react-compiler": "catalog:tooling",
+    "oxc-transform-react": "catalog:tooling",
     "racejar": "workspace:*",
     "react": "catalog:",
     "typescript": "catalog:tooling",
