+++ packages/plugin-sdk-value/package.json
@@ -71,7 +71,6 @@
     "@portabletext/editor": "workspace:^",
     "@portabletext/schema": "workspace:^",
     "@portabletext/test": "workspace:^",
-    "@rolldown/plugin-babel": "catalog:tooling",
     "@sanity/diff-match-patch": "catalog:",
     "@sanity/pkg-utils": "catalog:tooling",
     "@sanity/sdk-react": "^2.19.0",
@@ -82,8 +81,8 @@
     "@vitejs/plugin-react": "catalog:tooling",
     "@vitest/browser": "catalog:tooling",
     "@vitest/browser-playwright": "catalog:tooling",
-    "babel-plugin-react-compiler": "catalog:tooling",
     "jsdom": "catalog:tooling",
+    "oxc-transform-react": "catalog:tooling",
     "react": "catalog:",
     "react-dom": "catalog:",
     "typescript": "catalog:tooling",
