+++ packages/plugin-table/package.config.ts
@@ -2,7 +2,7 @@ import {defineConfig} from '@sanity/pkg-utils'
 
 export default defineConfig({
   tsconfig: 'tsconfig.dist.json',
-  reactCompiler: {target: '19'},
+  reactCompiler: {transform: 'oxc', target: '19'},
   // Lightning CSS lowers `light-dark()` into a `--lightningcss-light`/`-dark`
   // var pair that only resolves when it also processes the `color-scheme`
   // declaration. This stylesheet leaves `color-scheme` to the consumer, so the
