+++ packages/plugin-table/package.json
@@ -74,7 +74,6 @@
   "devDependencies": {
     "@portabletext/editor": "workspace:^",
     "@portabletext/test": "workspace:^",
-    "@rolldown/plugin-babel": "catalog:tooling",
     "@sanity/pkg-utils": "catalog:tooling",
     "@sanity/tsconfig": "catalog:tooling",
     "@tsdown/css": "catalog:tooling",
@@ -82,7 +81,7 @@
     "@vitejs/plugin-react": "catalog:tooling",
     "@vitest/browser": "catalog:tooling",
     "@vitest/browser-playwright": "catalog:tooling",
-    "babel-plugin-react-compiler": "catalog:tooling",
+    "oxc-transform-react": "catalog:tooling",
     "react": "catalog:",
     "react-dom": "catalog:",
     "typescript": "catalog:tooling",
